NLP6/requirements.txt
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.2.0
beautifulsoup4==4.12.3
cbor==1.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
datasets==2.19.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
executing==2.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
FlagEmbedding==1.3.3
frozenlist==1.5.0
fsspec==2024.3.1
gradio==5.8.0
gradio_client==1.5.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.3
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
inscriptis==2.5.0
ipython==8.30.0
ipywidgets==8.1.5
ir_datasets==0.5.9
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
langchain==0.3.9
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
litellm==1.53.7
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.57.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.13.2
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.6
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.2
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.2
trec-car-tools==2.6
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
unlzw3==0.2.2
urllib3==2.2.3
uvicorn==0.32.1
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
websockets==14.1
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zlib-state==0.1.9
